requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
nltk>=3.8.1
spacy>=3.7.0
wikipedia-api>=0.6.0
openai>=1.3.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-community>=0.0.10
pandas>=2.1.0
plotly>=5.17.0
reportlab>=4.0.0
markdown2>=2.4.0
python-docx>=1.1.0
python-dotenv>=1.0.0
wordcloud>=1.9.0
matplotlib>=3.8.0
sentence-transformers>=2.2.0
PyYAML>=6.0.1
groq>=0.4.0
